Fix オーバーフロー対策

+++ abc420/c/main.cpp
@@ -10,7 +10,7 @@ int main() {
     vector<int> A(N),B(N);
     for(int i=0;i<N;i++)cin>>A[i];
     for(int i=0;i<N;i++)cin>>B[i];
-    int total=0;
+    long long total=0;
     for(int i=0;i<N;i++)
     {
         total+=min(A[i],B[i]);
@@ -20,12 +20,11 @@ int main() {
         char c;cin>>c;
         int x,v;cin>>x>>v;
         x--;
-        total-=min(A[x],B[x]);
+        total-=(long long)min(A[x],B[x]);
         if(c=='A')A[x]=v;
         else if(c=='B')B[x]=v;
-        total+=min(A[x],B[x]);
+        total+=(long long)min(A[x],B[x]);
         cout<<total<<endl;
     }
     return 0;
 }
-
